pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "png2drawable"
version = "0.1.0"
description = "Desktop tool for generating Android drawable density variants from raster images"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "PNG2Drawable Contributors" }
]
keywords = ["android", "drawable", "qt", "pyside", "image", "desktop"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Environment :: X11 Applications :: Qt",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Operating System :: MacOS",
  "Operating System :: Microsoft :: Windows",
  "Operating System :: POSIX :: Linux",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Software Development :: Build Tools",
]
dependencies = [
  "Pillow>=11.2,<12",
  "PySide6>=6.9,<7",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3,<9",
  "ruff>=0.6,<1",
]

[project.scripts]
png2drawable = "main:main"

[tool.setuptools]
py-modules = ["main"]

[tool.setuptools.packages.find]
where = ["."]
include = ["core*", "gui*", "domain*", "services*", "infrastructure*"]
namespaces = true

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-q"
testpaths = ["tests"]
